Move connection mock tests to seperate test suite
