+++ mockrxandroidble/src/test/groovy/com/polidea/rxandroidble2/mockrxandroidble/RxBleClientMockTest.groovy
@@ -4,15 +4,9 @@ import android.os.Build
 import android.os.ParcelUuid
 import com.polidea.rxandroidble2.RxBleClient
 import com.polidea.rxandroidble2.RxBleConnection
-import com.polidea.rxandroidble2.exceptions.BleDisconnectedException
-import com.polidea.rxandroidble2.exceptions.BleGattCharacteristicException
-import com.polidea.rxandroidble2.exceptions.BleGattDescriptorException
-import com.polidea.rxandroidble2.mockrxandroidble.callbacks.results.RxBleGattReadResultMock
-import com.polidea.rxandroidble2.mockrxandroidble.callbacks.results.RxBleGattWriteResultMock
 import com.polidea.rxandroidble2.scan.ScanFilter;
 import com.polidea.rxandroidble2.scan.ScanSettings;
 import hkhc.electricspock.ElectricSpecification
-import io.reactivex.Observable
 import io.reactivex.subjects.PublishSubject
 import org.robolectric.annotation.Config
 
@@ -23,33 +17,10 @@ public class RxBleClientMockTest extends ElectricSpecification {
     def serviceUUID2 = UUID.fromString("00001235-0000-0000-8000-000000000000")
     def characteristicUUID = UUID.fromString("00002a29-0000-1000-8000-00805f9b34fb")
     def characteristicUUIDNoCallback = UUID.fromString("00002a29-0000-1000-8000-00805f9b34fc")
-    def characteristicNotifiedUUID = UUID.fromString("00002a29-0000-1000-8000-00805f9b34fb")
     def characteristicData = "Polidea".getBytes()
     def descriptorUUID = UUID.fromString("00001337-0000-1000-8000-00805f9b34fb")
     def descriptorData = "Config".getBytes()
     RxBleClient rxBleClient
-    PublishSubject characteristicNotificationSubject = PublishSubject.create()
-
-    PublishSubject<RxBleGattReadResultMock> characteristicReadSubject = PublishSubject.create()
-    PublishSubject<RxBleGattWriteResultMock> characteristicWriteSubject = PublishSubject.create()
-    PublishSubject<RxBleGattReadResultMock> descriptorReadSubject = PublishSubject.create()
-    PublishSubject<RxBleGattWriteResultMock> descriptorWriteSubject = PublishSubject.create()
-
-    def nextCharacteristicReadWillResult(closure) {
-        characteristicReadSubject.take(1).subscribe(closure)
-    }
-
-    def nextCharacteristicWriteWillResult(closure) {
-        characteristicWriteSubject.take(1).subscribe(closure)
-    }
-
-    def nextDescriptorReadWillResult(closure) {
-        descriptorReadSubject.take(1).subscribe(closure)
-    }
-
-    def nextDescriptorWriteWillResult(closure) {
-        descriptorWriteSubject.take(1).subscribe(closure)
-    }
 
     def createDevice(deviceName, macAddress, rssi) {
         new RxBleDeviceMock.Builder()
@@ -68,7 +39,6 @@ public class RxBleClientMockTest extends ElectricSpecification {
                 )
                 .connection(new RxBleConnectionMock.Builder()
                         .rssi(rssi)
-                        .notificationSource(characteristicNotifiedUUID, characteristicNotificationSubject)
                         .addService(serviceUUID,
                                 new RxBleClientMock.CharacteristicsBuilder()
                                         .addCharacteristic(
@@ -85,18 +55,6 @@ public class RxBleClientMockTest extends ElectricSpecification {
                                                         .build()
                                         ).build()
                         )
-                        .characteristicReadCallback(characteristicUUID, { device, characteristic, result ->
-                            characteristicReadSubject.onNext(result)
-                        })
-                        .characteristicWriteCallback(characteristicUUID, { device, characteristic, bytes, result ->
-                            characteristicWriteSubject.onNext(result)
-                        })
-                        .descriptorReadCallback(characteristicUUID, descriptorUUID, { device, descriptor, result ->
-                            descriptorReadSubject.onNext(result)
-                        })
-                        .descriptorWriteCallback(characteristicUUID, descriptorUUID, { device, descriptor, bytes, result ->
-                            descriptorWriteSubject.onNext(result)
-                        })
                         .build()
                 ).build()
     }
@@ -345,19 +303,6 @@ public class RxBleClientMockTest extends ElectricSpecification {
         testSubscriber.assertValue(42)
     }
 
-    def "should return the BluetoothDevice mtu"() {
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.requestMtu(72) }
-                .test()
-
-        then:
-        testSubscriber.assertValue(72)
-    }
-
     def "should return BluetoothDevices that were added on the fly"() {
         given:
         def discoverableDevicesSubject = PublishSubject.create()
@@ -377,286 +322,6 @@ public class RxBleClientMockTest extends ElectricSpecification {
         testNameSubscriber.assertValues("TestDevice", "SecondDevice")
         testAddressSubscriber.assertValues("AA:BB:CC:DD:EE:FF", "AA:BB:CC:DD:EE:00")
         testRssiSubscriber.assertValues(42, 17)
-    }
-
-    def "should return services list"() {
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection ->
-            rxBleConnection.discoverServices()
-                    .map { rxBleDeviceServices -> rxBleDeviceServices.getBluetoothGattServices() }
-                    .map { servicesList -> servicesList.size() }
-        }
-        .test()
-
-        then:
-        testSubscriber.assertValue(1)
-    }
-
-    def "should return characteristic data"() {
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readCharacteristic(characteristicUUIDNoCallback) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertValue("Polidea")
-    }
-
-    def "should return characteristic data via callback"() {
-        given:
-        nextCharacteristicReadWillResult { result -> result.success("PolideaCB".getBytes()) }
-        
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readCharacteristic(characteristicUUID) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertValue("PolideaCB")
-    }
-
-    def "should throw characteristic read error via callback"() {
-        given:
-        nextCharacteristicReadWillResult { result -> result.failure(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readCharacteristic(characteristicUUID) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertError { throwable ->
-            return BleGattCharacteristicException.class.isInstance(throwable) && ((BleGattCharacteristicException) throwable).getStatus() == 0x80
-        }
-    }
-
-    def "should throw characteristic read disconnection via callback"() {
-        given:
-        nextCharacteristicReadWillResult { result -> result.disconnect(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readCharacteristic(characteristicUUID) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertError { throwable ->
-            return BleDisconnectedException.class.isInstance(throwable) && ((BleDisconnectedException) throwable).state == 0x80
-        }
-    }
-
-    def "should write characteristic data via callback"() {
-        given:
-        nextCharacteristicWriteWillResult { result -> result.success() }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.writeCharacteristic(characteristicUUID, "PolideaCB".getBytes()) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertValue("PolideaCB")
-    }
-
-    def "should fail to write characteristic data via callback"() {
-        given:
-        nextCharacteristicWriteWillResult { result -> result.failure(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.writeCharacteristic(characteristicUUID, "PolideaCB".getBytes()) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertError({ throwable ->
-            return BleGattCharacteristicException.class.isInstance(throwable) && ((BleGattCharacteristicException) throwable).getStatus() == 0x80
-        })
-    }
-
-    def "should fail to write characteristic data due to disconnection via callback"() {
-        given:
-        nextCharacteristicWriteWillResult { result -> result.disconnect(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.writeCharacteristic(characteristicUUID, "PolideaCB".getBytes()) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertError({ throwable ->
-            return BleDisconnectedException.class.isInstance(throwable) && ((BleDisconnectedException) throwable).state == 0x80
-        })
-    }
-
-    def "should return descriptor data"() {
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readDescriptor(serviceUUID, characteristicUUIDNoCallback, descriptorUUID) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertValue("Config")
-    }
-
-    def "should return descriptor data via callback"() {
-        given:
-        nextDescriptorReadWillResult { result -> result.success("ConfigCB".getBytes()) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readDescriptor(serviceUUID, characteristicUUID, descriptorUUID) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertValue("ConfigCB")
-    }
-
-    def "should throw descriptor read error via callback"() {
-        given:
-        nextDescriptorReadWillResult { result -> result.failure(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readDescriptor(serviceUUID, characteristicUUID, descriptorUUID) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertError({ throwable ->
-            return BleGattDescriptorException.class.isInstance(throwable) && ((BleGattDescriptorException) throwable).getStatus() == 0x80
-        })
-    }
-
-    def "should throw descriptor read disconnection error via callback"() {
-        given:
-        nextDescriptorReadWillResult { result -> result.disconnect(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMapSingle { rxBleConnection -> rxBleConnection.readDescriptor(serviceUUID, characteristicUUID, descriptorUUID) }
-                .map { data -> new String(data) }
-                .test()
-
-        then:
-        testSubscriber.assertError({ throwable ->
-            return BleDisconnectedException.class.isInstance(throwable) && ((BleDisconnectedException) throwable).state == 0x80
-        })
-    }
-
-    def "should write descriptor data via callback"() {
-        given:
-        nextDescriptorWriteWillResult { result -> result.success() }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .take(1)
-                .flatMapCompletable { rxBleConnection -> rxBleConnection.writeDescriptor(serviceUUID, characteristicUUID, descriptorUUID, "ConfigCB".getBytes()) }
-                .test()
-
-        then:
-        testSubscriber.assertComplete();
-    }
-
-    def "should fail to write descriptor data via callback"() {
-        given:
-        nextDescriptorWriteWillResult { result -> result.failure(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .take(1)
-                .flatMapCompletable { rxBleConnection -> rxBleConnection.writeDescriptor(serviceUUID, characteristicUUID, descriptorUUID, "ConfigCB".getBytes()) }
-                .test()
-
-        then:
-        testSubscriber.assertError({ throwable ->
-            return BleGattDescriptorException.class.isInstance(throwable) && ((BleGattDescriptorException) throwable).getStatus() == 0x80
-        })
-    }
-
-    def "should fail to write descriptor data via callback due to disconnection"() {
-        given:
-        nextDescriptorWriteWillResult { result -> result.disconnect(0x80) }
-
-        when:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .take(1)
-                .flatMapCompletable { rxBleConnection -> rxBleConnection.writeDescriptor(serviceUUID, characteristicUUID, descriptorUUID, "ConfigCB".getBytes()) }
-                .test()
-
-        then:
-        testSubscriber.assertError({ throwable ->
-            return BleDisconnectedException.class.isInstance(throwable) && ((BleDisconnectedException) throwable).state == 0x80
-        })
-    }
-
-    def "should return notification data"() {
-        given:
-        def testSubscriber = rxBleClient.scanBleDevices(new ScanSettings.Builder().build())
-                .take(1)
-                .map { scanResult -> scanResult.getBleDevice() }
-                .flatMap { rxBleDevice -> rxBleDevice.establishConnection(false) }
-                .flatMap { rxBleConnection -> rxBleConnection.setupNotification(characteristicNotifiedUUID) }
-                .flatMap({ Observable<byte[]> observable -> observable })
-                .map { new String(it) }
-                .test()
-
-        when:
-        characteristicNotificationSubject.onNext("NotificationData".getBytes())
-
-        then:
-        testSubscriber.assertValue("NotificationData")
     }
 
     def "should emit correct connection state values when connected"() {
